Changed interaction logic for custom feedback and cases
Previous behaviour was that if the evaluation function produced custom feedback this feedback was returned instead of the matched case feedback. The new behaviour is that the evaluation functions custom feedback is appended to the case feedback, unless the case has the parameter "override_eval_feedback" set to True in which case the evaluation functions custom feedback is ignored.